requirements.txt
pydantic
python-dotenv
playwright
boto3
faker
requests
crewai
pyotp
